refactor(util): simplify CryptoRandomBytes parameter and improve test coverage (#86)
* refactor(util): simplify CryptoRandomBytes parameter and improve test coverage

- Change CryptoRandomBytes parameter type from int64 to int, removing unnecessary casts at call sites
- Migrate context_test.go and url_test.go to use testify assertions for consistency
- Add edge case tests for zero length, odd length, empty token/salt, and invalid URLs


* Update internal/util/url_test.go


* Update internal/util/url_test.go


* Update internal/util/crypto_test.go


---------

+++ internal/handlers/oauth_handler.go
@@ -249,7 +249,7 @@ func (h *OAuthHandler) OAuthCallback(c *gin.Context) {
 
 // generateRandomState returns a URL-safe base64-encoded string of nBytes random bytes.
 func generateRandomState(nBytes int) (string, error) {
-	b, err := util.CryptoRandomBytes(int64(nBytes))
+	b, err := util.CryptoRandomBytes(nBytes)
 	if err != nil {
 		return "", err
 	}
